+++ api/src/main/java/api/UserAnswerResource.java
@@ -2,9 +2,12 @@
 
  import api.auth.Auth;
  import rx.Observable;
- import se.fortnox.reactivewizard.jaxrs.PATCH;
 
- import javax.ws.rs.*;
+ import javax.ws.rs.Path;
+ import javax.ws.rs.PUT;
+ import javax.ws.rs.GET;
+ import javax.ws.rs.DELETE;
+ import javax.ws.rs.PathParam;
  import java.util.List;
 
 /**
