+++ impl/src/main/java/slack/ThreadMessageHandler.java
@@ -112,7 +112,7 @@ public class ThreadMessageHandler implements SlackMessageHandler {
                         question.setSlackId(mainMessageId);
                         question.setBounty(DEFAULT_BOUNTY);
 
-                        return first(questionResource.postQuestion(as(userId), question).doOnError(throwable -> LOG.error("Could not post message to slack", throwable))).thenReturn(question);
+                        return first(questionResource.createQuestion(as(userId), question).doOnError(throwable -> LOG.error("Could not post message to slack", throwable))).thenReturn(question);
                     }));
     }
